refactor: move diff request to single file
